support builtin iteration on itree
make for node in itree: possible. follow same rationale as append_child
and return wrapped ITreeNodes instead of the bare ITreeNode. the
traversal is breadth first since in the array it is a double for loop.
the outer loop is the level index and the inner loop is the sibling
index.

will need to add more functions to support other traversals