raise better error when accessing invalid nodes
accessing invalid nodes can mean that they don't exist, or that the
indices passed were of wrong type and could not be parsed.

when raising IndexError it wasn't clear what format the indices were
expected to have. this grouping needs further fine tuning at some point
however to allow distinguishing between different types of errors. it
is probably a bit too early for that though.