add transformer argument to from/to_nested_list
nested lists aren't easily modifyable from the outside, so it seemed
wise to add a way to modify them from the inside.

the transformer argument allows for internal manipulation of the
data before creating a tree and a nested list correspondingly.

in the case of from_nested_list, the callable gets passed an
element of the nested list as argument and the data attribute of the
created node is set to whatever the callable returns.

in the case of to_nested_list, the callable gets passed a
node of the tree as argument and the corresponding nested list element
is set to whatever the callable returns.

an example of why this is needed is json serialisation. the nested list
is already in json friendly format since it's a list of lists.
therefore, in order to create a json string representing the whole tree
we can turn the tree into a nested list of json strings and then
jsonize the nested list. in order to do that, notice that we can now
set the transformer to be: lambda x: node.to_json(x)